revert to benchmark % 10
